Set minimum redis-namespace version of 1.2.0
It looks like all previous versions of redis-namespace cause it to be
that the queues coming into MultiQueue have their names already prefixed
with the namespace. 1.2.0 breaks this by not having the name prefixed.
The patch from @mdimas at https://github.com/change/resque/commit/922c2e2ea5c44a949cf263ee94ae2369af635043
is not safe because it causes hash lookup misses on the latest
redis-namespace 1.2.0

Perhaps someone can look into why this is for backwards compatibility,
but I've got no interest in supporting old versions of redis-namespace
so this commit sets 1.2.0 as the minimum and fixes the errors with
MultiQueue.